Fix failure when no extensions are enabled
When the current extensions are none and the module attempts to get
the list of enabled extensions, the ast literal evaluation will fail.

This is because gsettings returns "@as []", for whatever reason. The
fix is an ugly but simple hack that uses a matching if statement. On
a match the returned value will instead be an empty list, as expected.