Fix set_extensions not working on other users
When ansible privilege escalation is used to become another user,
setting new values with gsettings does not work. This is because it
requires a running dbus session.

This is fixed by prefixing the existing set_extensions function with
the wonderful dbus-run-session command, which is made specifically
for oneshot command line operations, like this one.